Expand TortoiseORM demo, add README.md
Move the demo script to a dedicated subdirectory. Break out models and config into separate files. Add README
