+++ examples/tortoiseorm/config.py
@@ -1,0 +1,11 @@
+from tortoise import Tortoise
+
+TORTOISE_ORM = {
+    "connections": {"default": "postgres://arg@localhost:5432/tortoisedemo"},  # or your DB URL
+    "apps": {
+        "models": {
+            "models": ["models"],  # reference the models module
+            "default_connection": "default",
+        }
+    },
+}
