+++ examples/tortoiseorm/pyproject.toml
@@ -1,0 +1,4 @@
+[tool.aerich]
+tortoise_orm = "config.TORTOISE_ORM"
+location = "./migrations"
+src_folder = "./."
